# openapi.yaml
components:
  schemas:
...
    MemoryCreateRequest:
      additionalProperties: false
      description: "Body of ``POST /agent/memory`` \u2014 a user-added (manual) memory.\n\
        \n``user_id`` is intentionally absent: the subject is the verified bearer\n\
        identity, never client-supplied (cross-user-leak guard)."
      properties:
        content:
          maxLength: 2000
          minLength: 1
          title: Content
          type: string
        key:
          anyOf:
          - maxLength: 200
            type: string
          - type: 'null'
          default: null
          title: Key
        salience:
          anyOf:
          - maximum: 1.0
            minimum: 0.0
            type: number
          - type: 'null'
          default: null
          title: Salience
        type:
          default: semantic
          enum:
          - semantic
          - episodic
          - procedural
          title: Type
          type: string
      required:
      - content
      title: MemoryCreateRequest
      type: object
    MemoryItem:
      additionalProperties: false
      description: One stored memory item, surfaced to its owner in the panel.
      properties:
        content:
          title: Content
          type: string
        key:
          title: Key
          type: string
        salience:
          anyOf:
          - type: number
          - type: 'null'
          default: null
          title: Salience
        type:
          anyOf:
          - enum:
            - semantic
            - episodic
            - procedural
            type: string
          - type: 'null'
          default: null
          title: Type
      required:
      - key
      - content
      title: MemoryItem
      type: object
    MemoryListResponse:
      $defs:
        MemoryItem:
          additionalProperties: false
          description: One stored memory item, surfaced to its owner in the panel.
          properties:
            content:
              title: Content
              type: string
            key:
              title: Key
              type: string
            salience:
              anyOf:
              - type: number
              - type: 'null'
              default: null
              title: Salience
            type:
              anyOf:
              - enum:
                - semantic
                - episodic
                - procedural
                type: string
              - type: 'null'
              default: null
              title: Type
          required:
          - key
          - content
          title: MemoryItem
          type: object
      additionalProperties: false
      description: "Response body of ``GET /agent/memory`` \u2014 the caller's own\
        \ items."
      properties:
        items:
          items:
            $ref: '#/components/schemas/MemoryItem'
          title: Items
          type: array
      title: MemoryListResponse
      type: object
...
    MemorySuppressRequest:
      additionalProperties: false
      description: "Body of ``PATCH /agent/memory/{key}`` \u2014 soft-suppress / un-suppress\
        \ one\nof the caller's own memories (chat-persistence Phase B, D5).\n\n``suppressed=True``\
        \ (reject) stops the memory being recalled/injected while\nRETAINING the row\
        \ for audit; ``False`` un-suppresses (reversible). The\nsubject is the verified\
        \ bearer identity, never client-supplied."
      properties:
        suppressed:
          title: Suppressed
          type: boolean
      required:
      - suppressed
      title: MemorySuppressRequest
      type: object
...
    RunCreateRequest:
      additionalProperties: false
      description: Body of ``POST /agent/runs/stream`` and ``POST /agent/runs``.
      properties:
        agent_id:
          anyOf:
          - type: string
          - type: 'null'
          default: null
          title: Agent Id
        input:
          additionalProperties: true
          title: Input
          type: object
        thread_id:
          title: Thread Id
          type: string
      required:
      - thread_id
      - input
      title: RunCreateRequest
      type: object
...
    RunStateView:
      additionalProperties: false
      description: Response body of ``GET /agent/runs/{run_id}``.
      properties:
        completed_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Completed At
        run_id:
          title: Run Id
          type: string
        started_at:
          format: date-time
          title: Started At
          type: string
        status:
          enum:
          - running
          - completed
          - cancelled
          - errored
          title: Status
          type: string
        thread_id:
          title: Thread Id
          type: string
      required:
      - run_id
      - thread_id
      - status
      - started_at
      - completed_at
      title: RunStateView
      type: object
...
    ThreadListResponse:
      $defs:
        ThreadState:
          additionalProperties: false
          description: "Response body of ``GET /agent/threads/{thread_id}`` and items\
            \ in the\n``GET /agent/threads`` list.\n\n``title`` is the sidebar label\
            \ (Phase 3) \u2014 deterministic from the first\nuser message, user-renamable.\
            \ ``archived_at`` is the soft-delete tombstone\n(set on archive; list/get\
            \ filter it out)."
          properties:
            archived_at:
              anyOf:
              - format: date-time
                type: string
              - type: 'null'
              default: null
              title: Archived At
            created_at:
              format: date-time
              title: Created At
              type: string
            messages:
              items:
                additionalProperties: true
                type: object
              title: Messages
              type: array
            thread_id:
              title: Thread Id
              type: string
            title:
              default: New chat
              title: Title
              type: string
            updated_at:
              format: date-time
              title: Updated At
              type: string
            user_id:
              title: User Id
              type: string
          required:
          - thread_id
          - user_id
          - created_at
          - updated_at
          title: ThreadState
          type: object
      additionalProperties: false
      description: "Response body of ``GET /agent/threads`` \u2014 the caller's own\
        \ threads,\ncursor-paginated."
      properties:
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          default: null
          title: Next Cursor
        threads:
          items:
            $ref: '#/components/schemas/ThreadState'
          title: Threads
          type: array
      title: ThreadListResponse
      type: object
    ThreadRenameRequest:
      additionalProperties: false
      description: "Body of ``PATCH /agent/threads/{thread_id}`` \u2014 rename only."
      properties:
        title:
          maxLength: 200
          minLength: 1
          title: Title
          type: string
      required:
      - title
      title: ThreadRenameRequest
      type: object
...
info:
  description: AG-UI Agent Protocol routes exposed by agent_ui_adapter. Generated
    from Pydantic wire models.
  title: Agent UI Adapter
  version: 0.1.0
openapi: 3.1.0
paths:
  /agent/memory:
    get:
      operationId: listMemory
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryListResponse'
          description: The caller's memory items
        '401':
          description: Unauthorized
        '503':
          description: Memory service unavailable
      summary: List the caller's own memory items
    post:
      operationId: createMemory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryCreateRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryItem'
          description: Created memory
        '401':
          description: Unauthorized
        '503':
          description: Memory service unavailable
      summary: Add a user-authored memory
  /agent/memory/{key}:
    delete:
      operationId: deleteMemory
      parameters:
      - in: path
        name: key
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Deleted
        '401':
          description: Unauthorized
        '404':
          description: Memory not found
      summary: Delete one of the caller's memories
    patch:
      operationId: suppressMemory
      parameters:
      - in: path
        name: key
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemorySuppressRequest'
        required: true
      responses:
        '204':
          description: Suppression flag updated
        '401':
          description: Unauthorized
        '404':
          description: Memory not found
      summary: Soft-suppress / un-suppress one of the caller's memories (Phase B reject)
  /agent/runs/stream:
    post:
      operationId: streamRun
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunCreateRequest'
        required: true
      responses:
        '200':
          content:
            text/event-stream:
              schema:
                format: event-stream
                type: string
          description: AG-UI event stream
        '401':
          description: Unauthorized
      summary: Open an SSE stream of AG-UI events for a run
  /agent/runs/{run_id}:
    delete:
      operationId: cancelRun
      parameters:
      - in: path
        name: run_id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunStateView'
          description: Cancelled run state
        '404':
          description: Run not found
      summary: Cancel a run
    get:
      operationId: getRun
      parameters:
      - in: path
        name: run_id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunStateView'
          description: Run state
        '404':
          description: Run not found
      summary: Get run state
  /agent/threads:
    get:
      operationId: listThreads
      parameters:
      - in: query
        name: cursor
        required: false
        schema:
          type: string
      - in: query
        name: limit
        required: false
        schema:
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadListResponse'
          description: The caller's threads
        '401':
          description: Unauthorized
      summary: List the caller's own threads (newest-first, cursor)
    post:
      operationId: createThread
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ThreadCreateRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadState'
          description: Created thread
      summary: Create a thread
  /agent/threads/{thread_id}:
    delete:
      operationId: archiveThread
      parameters:
      - in: path
        name: thread_id
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Archived
        '401':
          description: Unauthorized
        '404':
          description: Thread not found
      summary: Archive (soft-delete) a thread
    get:
      operationId: getThread
      parameters:
      - in: path
        name: thread_id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadState'
          description: Thread state
      summary: Get a thread
    patch:
      operationId: renameThread
      parameters:
      - in: path
        name: thread_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ThreadRenameRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadState'
          description: Renamed thread
        '401':
          description: Unauthorized
        '404':
          description: Thread not found
      summary: Rename a thread
...